refactor: change package name for ApiUtil
